Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/preload/onload-event-expected.txt

CONSOLE MESSAGE: line 26: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: line 28: <link rel=preload> must have a valid `as` value

PASS Makes sure that preloaded resources trigger the onload event 
